feat: support serialization of pagination object
